Feat: udistrital/sga_documentacion#13 New model to verify recepits
