+++ controllers/formacion.go
@@ -6,7 +6,7 @@ import (
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
 	"github.com/udistrital/utils_oas/request"
-	"sga_mid_inscripcion/models"
+	"github.com/udistrital/sga_mid_inscripcion/models"
 	"strconv"
 	"strings"
 )
