+++ controllers/produccion_academica.go
@@ -7,7 +7,7 @@ import (
 	"github.com/astaxie/beego/logs"
 	"github.com/udistrital/utils_oas/request"
 	"github.com/udistrital/utils_oas/time_bogota"
-	"sga_mid_inscripcion/models"
+	"github.com/udistrital/sga_mid_inscripcion/models"
 )
 
 // ProduccionAcademicaController ...
