+++ controllers/solicitud_produccion.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/logs"
-	"sga_mid_inscripcion/models"
+	"github.com/udistrital/sga_mid_inscripcion/models"
 	"strconv"
 )
