+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	_ "sga_mid_inscripcion/routers"
+	_ "github.com/udistrital/sga_mid_inscripcion/routers"
 
 	apistatus "github.com/udistrital/utils_oas/apiStatusLib"
