+++ tests/default_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"runtime"
 	"path/filepath"
-	_ "sga_mid_inscripcion/routers"
+	_ "github.com/udistrital/sga_mid_inscripcion/routers"
 
 	"github.com/astaxie/beego"
 	. "github.com/smartystreets/goconvey/convey"
